Add button + Repair error
The error was:
A new entity was found through the relationship 'App\Entity\Doctor#schedules' that was not configured to cascade persist operations for entity: App\Entity\Schedule@461. To solve this issue: Either explicitly call EntityManager#persist() on this unknown entity or configure cascade persist this association in the mapping for example @ManyToOne(..,cascade={"persist"}). If you cannot find out which entity causes the problem implement 'App\Entity\Schedule#__toString()' to get a clue.